requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
aspose-words==22.5.0
atomicwrites==1.4.0
attrs==20.3.0
behave==1.2.6
better-profanity==0.7.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
decorator==4.4.2
et-xmlfile==1.1.0
excel2json==1.0.1
excel2json-3==0.1.6
Faker==8.1.0
helium==3.0.7
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.4.1
moviepy==1.0.3
msedge-selenium-tools==3.141.3
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandas==1.3.2
parse==1.19.0
parse-type==0.5.2
Pillow==8.2.0
pluggy==0.13.1
proglog==0.1.9
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.28.2
pytest==6.2.3
pytest-failed-screenshot==1.0.2
pytest-random-order==1.0.4
pytest-repeat==0.9.1
pytest-rerunfailures==10.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
Selenium-Screenshot==1.7.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.4
webdriver-manager==3.4.0
xlrd==1.2.0
XlsxWriter==3.0.1
